Refactor IssueDetail comments and tags
Refactored IssueDetail to use comments and tags fetch requests instead
of relying on the optional set(s) that were attached to an Issue.
This removed the use of `if let tags = issue.tags?.set as? Set<Tag>`.
This also allows sorting to be done in order to ensure same order when
view body updates.